add granular autosync controls
Enough of the Daktronics data feed has now been reverse engineered that
the score and penalties can be automatically synced in many cases.
Providing finer control over exactly what is synced is advantageous,
in case something goes wrong with the autosync code, the actual venue
scoreboard is wrong, or all of the needed data isn't in the sync feed
(e.g. loads of penalties and the scoreboard only shows the first 2).