start suppressing double calls to putTeamData()
These are when a single user action results in two or more calls to
putTeamData() in rapid succession. This is a bad situation, because
it could result in overwriting data on the server even though the
client side data is actually older. Error handling was added to catch
and abort these calls, and suppressTeamUpdate()/unsuppressTeamUpdate()
were added to allow explicit suppression and unsuppression. These two
functions are used by things like status button handlers to suppress
updates caused by changing other fields programmatically (is that even
necessary?)