Include period description in the data feed
This closes #29 for @vreelb. Detecting the end of the game or knowing whether
OT is sudden death was not actually a blocker here. The current code just
assumes that the game will end if we are in the last period or an overtime
period and the scores are unequal, or if we are in the last allowed overtime
period. This logic may not be correct for all sports (maybe soccer?)