Implement Daktronics RTD penalty sync
This implementation is a bit crude in that it bypasses the internal
penalty queue logic. This means that the penalty data from the RTD
feed will not appear in the user interface. If the RTD penalty data
is wrong or out of date, turning off "other" sync will disable
display of the penalty data acquired from the sync feed.